requirements.txt
Flask==2.0.2
mysqlclient==2.0.3
pymysql==1.2.5
Flask-MySQLdb==0.2.0

